requirements.txt
flask==3.0.0
cryptography==41.0.5
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
Flask-WTF==1.2.1
gunicorn==20.1.0
flask-sqlalchemy==3.0.5
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.70.0
pandas==1.5.3